+++ src/main/cpp/it/d4np/util/util.hpp
@@ -8,6 +8,7 @@
 #define IT_D4NP_UTIL_UTIL_HPP
 
 #include <it/d4np/util/circular_buffer.hpp>
+#include <it/d4np/util/cli_parser.hpp>
 #include <it/d4np/util/flat_map.hpp>
 #include <it/d4np/util/flat_set.hpp>
 #include <it/d4np/util/hash.hpp>